// src/app/models/warehouse/warehouse-model.ts
export class WarehouseModel {
    id: string;
    name: string;
    addressDetail: string;
    districtName: string;
    districtId: number;
    email: string;
    provinceId: number;
    provinceName: string;
    roadId: number;
    roadName: string;
    status: number;
    phoneNumber: string;
    vendorId: string;
    warehouseCode: string;
    warehouseType: number;
    warehouseDescription: string;
    statusName: string;
    typeName: string;
    vendorName: string;
    villageId: number;
    villageName: string;

}